+++ src/main/java/mk/edu/uklo/fikt/fiktexamweb/util/TopicRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 @Repository
-public interface TopicRepository extends JpaRepository<Topic, Long>{
-    List<Topic> findBySubjectId(long subjectId);
+public interface TopicRepository extends JpaRepository<Topic, Integer>{
+    List<Topic> findBySubjectId(int subjectId);
 
 }
